translate(dialog): rename DIALOG to editor_dialog
