+++ editor_dialog.cpp
